refactor: analytics.framework
closes https://github.com/NathanWalker/angular2-seed-advanced/issues/34
