+++ nativescript/app/pages/app/app.component.ts
@@ -11,7 +11,8 @@ import {RouterState} from 'ngrx-store-router';
 
 // app
 import {AppComponent} from '../../components/app/app.component';
-import {LogService, AnalyticsService} from '../../frameworks/core.framework/index';
+import {LogService} from '../../frameworks/core.framework/index';
+import {AnalyticsService} from '../../frameworks/analytics.framework/index';
 import {ModalNative} from '../../shared/core/services/modal-native.service';
 import {ActionBarUtil} from '../../shared/core/utils/actionbar.util';
