+++ src/frameworks/core.framework/index.ts
@@ -2,29 +2,14 @@
 import {HTTP_PROVIDERS} from 'angular2/http';
 import {ROUTER_PROVIDERS} from 'angular2/router';
 
-// libs
-import {Angulartics2} from 'angulartics2';
-import {Angulartics2Segment} from 'angulartics2/src/providers/angulartics2-segment';
-
 // app
-import {AnalyticsService} from './services/analytics.service';
 import {LogService} from './services/log.service';
-
-export const ANALYTICS_PROVIDERS: any[] = [
-  Angulartics2,
-  Angulartics2Segment,
-  AnalyticsService
-];
 
 export const CORE_PROVIDERS: any[] = [
   HTTP_PROVIDERS,
   ROUTER_PROVIDERS,
-  LogService,
-  ANALYTICS_PROVIDERS
+  LogService
 ];
-
-// services
-export * from './services/analytics.service';
 
 // decorators
 export * from './decorators/base.component';
@@ -32,7 +17,6 @@ export * from './decorators/form.component';
 export * from './decorators/route.component';
 
 // interfaces
-export * from './interfaces/ianalytics';
 export * from './interfaces/iconsole';
 export * from './interfaces/iwindow';
 export * from './interfaces/ilang';
