--- src/frameworks/core.framework/interfaces/ianalytics.ts
@@ -1,9 +1,0 @@
-export interface IAnalyticsProperties {
-  category?: string;
-  label?: string;
-  value?: number;
-}
-
-export interface IAnalytics {
-  track(action: string, properties: IAnalyticsProperties): void;
-}
